+++ Engine/Source/RHI-Dummy/Include/RHI/Dummy/CommandRecorder.h
@@ -87,7 +87,7 @@ namespace RHI::Dummy {
         void SetIndexBuffer(BufferView* bufferView) override;
         void SetVertexBuffer(size_t slot, BufferView* bufferView) override;
         void Draw(size_t vertexCount, size_t instanceCount, size_t firstVertex, size_t firstInstance) override;
-        void DrawIndexed(size_t indexCount, size_t instanceCount, size_t firstIndex, size_t baseVertex, size_t firstInstance) override;
+        void DrawIndexed(size_t indexCount, size_t instanceCount, size_t firstIndex, int32_t baseVertex, size_t firstInstance) override;
         void SetViewport(float topLeftX, float topLeftY, float width, float height, float minDepth, float maxDepth) override;
         void SetScissor(uint32_t left, uint32_t top, uint32_t right, uint32_t bottom) override;
         void SetPrimitiveTopology(PrimitiveTopology primitiveTopology) override;
